requirements.txt
Flask_Login==0.5.0
Flask_Bcrypt==0.7.1
Flask_SQLAlchemy==2.4.4
Flask==1.1.2
Flask_WTF==0.14.3
WTForms==2.3.3
mysql_connector_repackaged==0.3.1
